+++ src/index.tsx
@@ -3,14 +3,14 @@ import { Spin } from 'antd';
 import { Provider } from 'mobx-react';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { globalStore } from './models/global';
+import { globalModel } from './models/globalModel';
 import { routes } from './routes';
 import { Locale, RenderRoutes, browserHistory } from './utils';
 
 ReactDOM.render(
   <React.StrictMode>
     <Locale>
-      <Provider globalStore={globalStore}>
+      <Provider globalModel={globalModel}>
         <RenderRoutes
           history={browserHistory}
           loading={<Spin size="large"/>}