+++ src/utils/renderRoutes.tsx
@@ -3,63 +3,17 @@ import React from 'react';
 import { Redirect, Router, Route, Switch } from 'react-router-dom';
 
 export interface MenuDataItem {
-
-  /**
-   * @name 子菜单
-   */
+  Component?: React.ComponentType;
+  Wrapper?: React.ComponentType;
+  exact?: boolean;
+  sensitive?: boolean;
+  strict?: boolean;
+  redirect?: string;
   children?: MenuDataItem[];
-
-  /**
-   * @name 在菜单中隐藏子节点
-   */
-  hideChildrenInMenu?: boolean;
-
-  /**
-   * @name 在菜单中隐藏自己和子节点
-   */
-  hideInMenu?: boolean;
-
-  /**
-   * @name 菜单的icon
-   */
   icon?: React.ReactNode;
-
-  /**
-   * @name 自定义菜单的国际化 key
-   */
-  locale?: string | false;
-
-  /**
-   * @name 菜单的名字
-   */
   name?: string;
-
-  /**
-   * @name 用于标定选中的值，默认是 path
-   */
-  key?: string;
-
-  /**
-   * @name disable 菜单选项
-   */
-  disabled?: boolean;
-
-  /**
-   * @name 路径
-   */
   path?: string;
-
-  /**
-   * @name 自定义父节点
-   * @description 当此节点被选中的时候也会选中 parentKeys 的节点
-   */
-  parentKeys?: string[];
-
-  /**
-   * @name 隐藏自己，并且将子节点提升到与自己平级
-   */
-  flatMenu?: boolean;
-  [key: string]: any;
+  key?: string;
 }
 
 const DefaultWrapper: React.FC<any> = function(props) {
@@ -125,9 +79,9 @@ function renderRoutes(routes: MenuDataItem[], extraProps = {}, switchProps = {},
                 loading={loading}
                 path={route.path}
                 render={(props) => {
-                  const childRoutes = renderRoutes(route.routes, extraProps, { location: props.location }, loading);
-                  if (route.component) {
-                    const { component: Component } = route;
+                  const childRoutes = renderRoutes(route.children, extraProps, { location: props.location }, loading);
+                  if (route.Component) {
+                    const { Component } = route;
                     return (
                       <Component
                         {...props}